fixed tests, clippy and rust size
